requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
py-dotenv==0.1
python-dotenv==0.10.1
requests==2.21.0
urllib3==1.24.2
